// src/types/form.ts
export type loginValidation = {
  email?: string | null;
  password?: string | null;
};
export type registerValidation = {
  name?: string | null;
  email?: string | null;
  passWord?: string | null;
  cpassword?: string | null;
};
export type registertype = {
  name: string;
  email: string;
  passWord: string;
  cpassword: string;
};